+++ client/src/styles/CardModal.scss
@@ -10,79 +10,83 @@
 
 * {
   font-family: var(--font-family-1);
+  box-sizing: border-box;
+  padding: 0;
+  margin: 0;
 }
 
-@media screen and (min-width: 992px) {
-  .modal-container {
-    padding: 80px;
-    background: linear-gradient(to right, var(--color-5), var(--color-2));
-  }
+.modal-container {
+  width: 100vw;
+  height: 100vh;
+  position: absolute;
+  left: 0;
+  top: -190px;
+  margin: 0;
+}
+
+.modal {
+  border-radius: 10px;
+  color: black;
+  z-index: 9999;
 }
 
 .modal-header {
-  background-image: url("src/assets/images/Mexico.jpg");
-  background-size: cover;
-  background-repeat: no-repeat;
-  padding-bottom: 60vw;
+  height: 35vh;
 }
 
-.arrow-container {
-  display: flex;
-  justify-content: flex-start;
+.modal-image {
+  width: 100%;
+  height: 100%;
 }
 
 .arrow {
   cursor: pointer;
+  position: relative;
+  width: 2rem;
+  top: -310px;
+  left: 390px;
 }
 
 .modal-content {
   background-color: white;
-  border-radius: 20px 20px 0 0;
-  margin-top: -15px;
-}
-
-.modal-content-header {
-  margin: auto 5vw auto 5vw;
+  height: 65vh;
+  box-sizing: border-box;
+  padding: 1rem 1.5rem;
+  overflow-y: auto;
 }
 
 .modal-content-header-heart {
   display: flex;
   justify-content: space-between;
-  margin-bottom: -30px;
 }
 
 .countryH1 {
   color: var(--color-3);
-}
-
-.heart {
-  cursor: pointer;
-  opacity: 0.6;
+  font-size: 1.8rem;
 }
 
 .modal-content-weather {
-  margin: auto 5vw auto 5vw;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
 }
 
 .weather-first-block {
   display: flex;
   flex-direction: column;
   align-items: center;
-  gap: 0px;
 }
 
-.weather-info {
-  font-weight: bold;
-  font-size: 1.2rem;
+.meteo-icon {
+  width: 50px;
+  height: auto;
 }
 
 .weather-second-block {
   display: flex;
-  justify-content: space-around;
-}
-
-.modal-content-converter {
-  margin: auto 5vw auto 5vw;
+  flex-direction: row;
+  align-items: center;
+  gap: 1rem;
 }
 
 .button-convert {
@@ -100,6 +104,9 @@
   font-size: 1.1rem;
 }
 
-.modal-content-holidays {
-  margin: auto 5vw auto 5vw;
-}
+/*@media screen and (min-width: 992px) {
+  .modal-container {
+    padding: 80px;
+    background: linear-gradient(to right, var(--color-5), var(--color-2));
+  }
+}*/
