+++ client/src/styles/RecommandedSection.scss
@@ -21,6 +21,7 @@ body {
 h3 {
   font-family: $primary-font;
   font-size: 1.5rem;
+  color: white;
 }
 
 .white {
@@ -31,7 +32,6 @@ h3 {
 .recommended__section {
   position: relative;
   top: -70px;
-  margin: 0 2rem;
   .recommended__carousel {
     display: flex;
     gap: 1.5rem;
