+++ client/src/styles/SectionPopular.scss
@@ -22,7 +22,6 @@ body {
 h3 {
   font-family: $primary-font;
   font-size: 1.5rem;
-  color: black;
 }
 
 h4 {
